requirements.txt
llama-index
python-dotenv
jupyter